Window functions: Sql.Window.Rank must return long, not int
RANK() is bigint (or wider) on every supported backend, and the sibling
window ranking functions (RowNumber/DenseRank/NTile) and the legacy
Sql.Ext.Rank() all return long. WindowFunctionBuilder becomes public in
this PR, so finalize the return type now (non-breaking before exposure).

This also removes the int->long reconciliation cast the legacy-analytic
conversion was forcing on Rank, which split the value into two window
columns in a subquery (TestSubqueryOptimization emitted RANK() twice).

+++ Source/LinqToDB/Sql/Sql.Window.cs
@@ -559,7 +559,7 @@ namespace LinqToDB
 		/// FROM Table t
 		/// </code>
 		/// </remarks>
-		public static int Rank(this Sql.IWindowFunction window, Func<IOPartitionROrderFinal, IDefinedFunction> func)
+		public static long Rank(this Sql.IWindowFunction window, Func<IOPartitionROrderFinal, IDefinedFunction> func)
 			=> throw new ServerSideOnlyException(nameof(Rank));
 
 		/// <summary>
